fix(ui): move `getItemsPerRow` to frontend src dir
Not sure how but it was in repo root

Closes #8509
