+++ invokeai/frontend/web/src/features/gallery/components/getItemsPerRow.ts
